fix work with flyway

+++ src/main/resources/db/migration/V1__init_db.sql
@@ -1,18 +1,18 @@
 create sequence hibernate_sequence start 1 increment 1;
 
 create table user_messages (
-	id int8 not null, 
+	id bigint not null,
 	create_at timestamp, 
 	text varchar(255) not null, 
-	user_id int8, primary key (id));
+	user_id bigint, primary key (id));
 
 create table user_role (
-	user_id int8 not null, 
+	user_id bigint not null,
 	roles varchar(255)
 );
 
 create table usr (
-	id int8 not null, 
+	id bigint not null,
 	activation_code varchar(255), 
 	active boolean not null, 
 	email varchar(64) not null,
